+++ src/layman/__init__.py
@@ -34,8 +34,10 @@ for bp in get_blueprints():
     app.register_blueprint(bp, url_prefix='/rest/<username>')
 
 from .user.rest_current_user import bp as current_user_bp
+from .gs_wfs_proxy import bp as gs_wfs_proxy_bp
 
 app.register_blueprint(current_user_bp, url_prefix='/rest/current-user')
+app.register_blueprint(gs_wfs_proxy_bp, url_prefix='/geoserver')
 
 app.logger.info(f"IN_CELERY_WORKER_PROCESS={IN_CELERY_WORKER_PROCESS}")
 app.logger.info(f"IN_PYTEST_PROCESS={IN_PYTEST_PROCESS}")